+++ engine/src/main/kotlin/engine/context/Preset.kt
@@ -26,11 +26,24 @@ enum class Preset(val description: String, val settings: Map<Setting, SettingVal
         "A set of settings which make the solutions compatible with Graspable Math.",
         mapOf(
             Setting.DontAddClarifyingBrackets setTo BooleanSetting.True,
+            Setting.BalancingMode setTo BalancingModeSetting.NextTo,
+            Setting.MoveTermsOneByOne setTo BooleanSetting.True,
+            Setting.QuickAddLikeFraction setTo BooleanSetting.True,
+            Setting.QuickAddLikeTerms setTo BooleanSetting.True,
+            Setting.CommutativeReorderInSteps setTo BooleanSetting.True,
+            Setting.CopySumSignsWhenDistributing setTo BooleanSetting.True,
+        ),
+    ),
+
+    GMFriendlyAdvanced(
+        "A set of settings which make the solutions compatible with Graspable Math.",
+        mapOf(
+            Setting.DontAddClarifyingBrackets setTo BooleanSetting.True,
             Setting.BalancingMode setTo BalancingModeSetting.Advanced,
             Setting.MoveTermsOneByOne setTo BooleanSetting.True,
             Setting.QuickAddLikeFraction setTo BooleanSetting.True,
             Setting.QuickAddLikeTerms setTo BooleanSetting.True,
-            Setting.ReorderProductsInSteps setTo BooleanSetting.True,
+            Setting.CommutativeReorderInSteps setTo BooleanSetting.True,
             Setting.CopySumSignsWhenDistributing setTo BooleanSetting.True,
         ),
     ),